+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [289](https://github.com/vegaprotocol/data-node/pull/289) - Add support for pagination of delegations
 - [254](https://github.com/vegaprotocol/data-node/pull/254) - Move to `ghcr.io` container registry
 - [290](https://github.com/vegaprotocol/data-node/pull/290) - Update pegged orders offset
+- [296](https://github.com/vegaprotocol/data-node/pull/296) - Expose validator performance score attributes on Node object 
 
 
 ### 🐛 Fixes